Refactor routes to make more sense
